add diff of Rop w.r.t. Temp and test

+++ pyjac/tests/__init__.py
@@ -87,7 +87,8 @@ class storage(object):
 
         # various indicies and mappings
         self.rev_inds = np.array(
-            [i for i in range(gas.n_reactions) if gas.is_reversible(i)])
+            [i for i in range(gas.n_reactions) if gas.is_reversible(i)],
+            dtype=np.int32)
         self.rev_rate_constants = np.zeros((test_size, self.rev_inds.size))
         self.rev_rxn_rate = np.zeros((test_size, self.rev_inds.size))
         self.equilibrium_constants = np.zeros((test_size, self.rev_inds.size))
